move endpoint to pkg
