+++ server/schedule/labeler/labeler.go
@@ -22,10 +22,10 @@ import (
 
 	"github.com/pingcap/log"
 	"github.com/tikv/pd/pkg/errs"
+	"github.com/tikv/pd/pkg/storage/endpoint"
 	"github.com/tikv/pd/pkg/utils/syncutil"
 	"github.com/tikv/pd/server/core"
 	"github.com/tikv/pd/server/schedule/rangelist"
-	"github.com/tikv/pd/server/storage/endpoint"
 	"go.uber.org/zap"
 )
